backend/requirements.txt
motor
fastapi
uvicorn
pydantic
bson
python-dotenv
bcrypt
cloudinary
python-multipart
aiohttp